Move studyDataUpdate to study.js vejkkkk
