+++ app/scripts/redact.js
@@ -1,5 +1,5 @@
 import { DateTime } from "luxon";
-import { mapreplacer, mapreviver } from "./utils.js";
+import { updateStudyData } from "./study.js";
 
 const dateprevFormat = "yyyy-MM-dd HH:mm:ss";
 
@@ -221,36 +221,6 @@ function getTimestampType(el) {
           updateStudyData(tsType, msu);
         }
       });
-    }
-    function updateStudyData(tsType, msu) {
-      // increment counter in local storage area
-      browser.storage.local
-        .get("msuChoices") // {TSTYPE: {MSU: Counter, ...}, ...}
-        .then((res) => {
-          var msuChoices = res.msuChoices;
-          if (msuChoices === undefined) {
-            msuChoices = new Map();
-          } else {
-            msuChoices = JSON.parse(msuChoices, mapreviver);
-          }
-
-          var tsStats = msuChoices.get(tsType);
-          if (tsStats === undefined) {
-            tsStats = new Map();
-          }
-          var tsMsuCount = tsStats.get(msu);
-          if (tsMsuCount === undefined) {
-            tsMsuCount = 0;
-          }
-          tsMsuCount++;
-          tsStats.set(msu, tsMsuCount);
-          msuChoices.set(tsType, tsStats);
-          let stringified = JSON.stringify(msuChoices, mapreplacer);
-          console.log(`LS: msuChoices ${stringified}`);
-
-          // store updated stats
-          browser.storage.local.set({ msuChoices: stringified });
-        });
     }
 
     function redactTimestamps() {
